fixed front end
fixed front end base.html

+++ wallet/models.py
@@ -52,7 +52,7 @@ class UserProfileInfo(models.Model):
     user = models.OneToOneField(settings.AUTH_USER_MODEL, on_delete=models.CASCADE, related_name='profile')
     first_name = models.CharField(max_length=100, null=True)
     last_name = models.CharField(max_length=100, null=True)
-    coin_balance = models.DecimalField(max_digits=10, decimal_places=2, default=0.0)
+    coin_balance = models.FloatField(default=0.0)
     point_balance = models.FloatField(default=0.0)
     points_to_give = models.FloatField(default=0.0)
     profile_picture = models.ImageField(upload_to='profile_pics/', null=True, blank=True)
